Moved prombench/cmd to /infra
